LayoutTests/platform/ios-wk2/imported/w3c/web-platform-tests/css/css-flexbox/radiobutton-min-size-expected.txt

You should see two identical-looking lines, both with a radio button at the beginning.


Text
Text

FAIL two radio button widths are identical assert_equals: width should be equal expected 0 but got 16
